The End Product

+++ src/me/venom/superrant/interfaces/IRentalPricing.java
@@ -2,6 +2,6 @@ package me.venom.superrant.interfaces;
 
 public interface IRentalPricing
 {
-    float getRentalPrice();
-    float getOverdueFee();
+    double getRentalPrice();
+    double getOverdueFee();
 }
